TP2/requirements.txt
# Dependencias del TP2 - Sistema de Scraping y Análisis Web Distribuido

# Cliente HTTP asíncrono
aiohttp>=3.9.0
aiofiles>=23.0.0

# Parsing HTML (sin lxml para evitar problemas de compilación en Windows)
beautifulsoup4>=4.12.0

# Procesamiento de imágenes
Pillow>=10.0.0

# Web scraping y screenshots
selenium>=4.15.0
requests>=2.31.0

# Testing (opcional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
